+++ shared/budgie-components/src/lib/all-transactions-view/AllTransactionsView.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import AllTransactionsView from './AllTransactionsView';
+
+describe('AllTransactionsView', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<AllTransactionsView />);
+    expect(baseElement).toBeTruthy();
+  });
+});
