+++ shared/budgie-components/src/lib/all-transactions-view/AllTransactionsView.stories.js
@@ -1,0 +1,12 @@
+import { AllTransactionsView } from './AllTransactionsView';
+
+export default {
+  component: AllTransactionsView,
+  parameters: {
+    layout: 'fullscreen',
+  },
+};
+
+export const DefaultAllTransactionsView = {
+  render: () => <AllTransactionsView />, // this function specifies how to render a variant
+};
