+++ shared/budgie-components/src/lib/monthly-transactions-view/MonthlyTransactionsView.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import MonthlyTransactionsView from './MonthlyTransactionsView';
+
+describe('MonthlyTransactionsView', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<MonthlyTransactionsView />);
+    expect(baseElement).toBeTruthy();
+  });
+});
