+++ shared/budgie-components/src/lib/monthly-transactions-view/MonthlyTransactionsView.stories.js
@@ -1,0 +1,12 @@
+import { MonthlyTransactionsView } from './MonthlyTransactionsView';
+
+export default {
+  component: MonthlyTransactionsView,
+  parameters: {
+    layout: 'fullscreen',
+  },
+};
+
+export const DefaultMonthlyTransactionsView = {
+  render: () => <MonthlyTransactionsView />, // this function specifies how to render a variant
+};
